Sent mail
This script will require you to pass in the email id and the password, and if you don't want less secure apps to access your Google account you have to enable 2 step verification.

Which is a bit of a tedious process and also for this script to work you also might want to disable the firewall.

So I will make a generalized program that will alert you when the prices have dropped instead of sending a mail.

Anyway, this is the script that will let you know whether the prices have dropped!