Working with GUI...
Just started to learn about adding a gui for my program using python's tkinter.

I just commented out my driver code to just check whether the basic interface is working properly.
The code is not complete, but it will show you the product name and price if you enter the link and the desired price onto the slots provided in the interface. That's only for now, more updates soon....