Better Exception Handling
Situations like 'no internet connection' and different user agents for different users could be handled now.

Updates made:
- User agent input: I should not force the user to edit the code and add his user-agent right? Therefore whenever a new user opens the application and clicks any of the the buttons in the interface, a new window will open up asking for the user agent. The window will have enough information on how the user could find the user-agent.

Even if the user closes the window (don't be that rude) without entering anything or types something else, an error message will be shown and the user will be prompted again to enter the user-agent.

It is highly recommended that you should copy and paste the entire user-agent. If you miss a couple of letters in the end, the app will accept the entry. Even though the application will be able to grab information from the webpage, it might not happen always.

So please copy and paste the ENTIRE USER AGENT.

I have used the module httpagentparser to check whether the user agent is valid.

- No internet error: For the app to work, you must be connected to the internet. If not an error message will be displayed.

- A check all and reset all button is added. (It is misaligned as for now) Reset or Check all fields using those buttons.
Check all option will take some time to display all the information, if you have all the input fields filled. So you have to wait for the app to display the info.