Visualizaton of price fluctuations
Using Matplotlib and Pandas, I added an extra feature to this application.

Every time you check the price of a product, the current price of that product will be stored into the text file. With that data, a line graph will be created, which shows the price variations that occured for the product.

Using this plot, you will be able to predict how often the price of your desired product varies. Hence this acts as a really good tool for analysis.

Remember, a straight line indicates that there was no price fluctuation whenever you checked for the price using the app.

Recent 20 values will be stored for analysis. When newer values are added, the old ones get deleted.

Currently in this script, this functionality is added only for Product1. Click the launch button to view the plot. UI improvements will be made soon.