COMPLETE CODE: DAY 15
Project completed successfully!

This is the complete program of the project ' Track prices of products in Amazon.in'.

Today's update: Confirmation of Reset all button - When Reset all button is clicked, a message box pops up, asking us to confirm the action, for preventing the user from clicking the button accidentally.

 Remember, a text file named pricetracker.txt will be created in the same location as that of this python file, as database for the application.

*** About check all button: If you have all the entry fields filled with product urls and hits the check all button, it will take some time to fetch the information. So kindly wait for sometime and do not close the application. The results will be displayed after some time. ***

Before running this code, ensure that all the imported packages and modules are installed into the system.

Packages required are:
 - bs4
- requests
- httpagentparser
- pandas
- matplotlib
- numpy
(use pip install )

All the other packages are available in the standard library. So you need not install those.

All right, enjoy!